Enum example_module::example_module::GitSignatureState

pub enum GitSignatureState {
    BAD_EMAIL,
    EXPIRED_KEY,
    GPGVERIFY_ERROR,
    GPGVERIFY_UNAVAILABLE,
    INVALID,
    MALFORMED_SIG,
    NOT_SIGNING_KEY,
    NO_USER,
    OCSP_ERROR,
    OCSP_PENDING,
    OCSP_REVOKED,
    UNKNOWN_KEY,
    UNKNOWN_SIG_TYPE,
    UNSIGNED,
    UNVERIFIED_EMAIL,
    VALID,
    Other(String),
}

Variants
BAD_EMAIL

Invalid email used for signing.

EXPIRED_KEY

Signing key expired.

GPGVERIFY_ERROR

Internal error - the GPG verification service misbehaved.

GPGVERIFY_UNAVAILABLE

Internal error - the GPG verification service is unavailable at the moment.

INVALID

Invalid signature.

MALFORMED_SIG

Malformed signature.

NOT_SIGNING_KEY

The usage flags for the key that signed this don't allow signing.

NO_USER

Email used for signing not known to GitHub.

OCSP_ERROR

Valid siganture, though certificate revocation check failed

OCSP_PENDING

Valid signature, pending certificate revocation checking

OCSP_REVOKED

One or more certificates in chain has been revoked

UNKNOWN_KEY

Key used for signing not known to GitHub.

UNKNOWN_SIG_TYPE

Unknown signature type.

UNSIGNED

Unsigned.

UNVERIFIED_EMAIL

Email used for signing unverified on GitHub.

VALID

Valid signature and verified by GitHub.

Other(String)